library/book_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
uuid
PyJWT
datetime
pydantic[email]
alembic
netcat
grpcio==1.66.1
grpcio-tools==1.66.1